Refactor separate button component into nested component
