+++ components/philips_series_2200/__init__.py
@@ -6,11 +6,9 @@ from esphome.const import CONF_ID
 
 AUTO_LOAD = [
     "switch",
-    "button",
     "text_sensor",
     "number",
     "philips_power_switch",
-    "philips_action_button",
     "philips_status_sensor",
     "philips_bean_settings",
     "philips_size_settings",
