+++ components/philips_series_2200/button/action_button.cpp
